[SYCL][OpenMP] Add support to remove temporary files generated by unbundler
When initial support of offloading with fat static archives was implemented,
we did not have a viable solution to remove the temporary files generated
from the unbundler.  These changes address that problem by creating a new
file type that is expected to be a list of files which can be removed.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>