[SYCL] Redesign memory allocation in buffer_impl class.
Do not allocate memory on host in buffer_impl constructors when user's
ptr is provided. This is possible because the ownership of that memory
is given to the constructed SYCL buffer for the duration of it'
lifetime. The memory still need to be allocated in constructor that
takes iterators as they can point to non-consecutive memory(need
consecutive memory to interact with OpenCL).
This may improve performance as we avoid several memory allocation/copy.

As required by the Specification the host memory now always allocated
using allocator provided by user.

Moved methods that allocate, fill and copy memory to separate class.
Copy method was improved to support device to host and host to device
partial(aka CopyRect) version of copies. So, now the buffer_impl class
is decoupled from OpenCL.

Fix possible bug, now constructor from shared_ptr stores a copy of
shared_ptr in lambda to avoid deallocation of memory until buffer is
destructed or set_final_data is invoked to override destination for
updating host memory(aka copy back).

Signed-off-by: Vlad Romanov <vlad.romanov@intel.com>