[SYCL] Introduction of new scheduler.
This patch introduces new scheduler which is enabled when SCHEDULER_20
macro is set(set by default). The new scheduler is based on accessor
rather on buffer, so it will support images.

Also now there are two commands that implements moving memory to
other contexts instead of one - alloca memory and memcpy. There is new
command - release memory which will be ran during the sycl::buffer or
sycl::image object destruction and releases memory instances for the
memory object.

Signed-off-by: Vlad Romanov <vlad.romanov@intel.com>