[SYCL] Remove wrapper headers
Even when compiling for the device, code meant only for host execution
must be semantically analyzable. Existing wrapper headers were initially
created to resolve this problem. But now -aux-triple is used to pass
target triple of opposite side of compilation. Clang knows target triple
for both sides of compilation so that preprocessor macros and target
builtins from both sides are available. That is why there is no need in
existing wrapper headers with stub definitions since they break
compilation for cases when standard headers are included.

Approach with -aux-triple is not complete solution. Host headers might
provide definitions relying on host inline asm, intrinsics, using types
not lowerable in device code. So we will need to provide a solution to
overload some definitions for the device or to define only some macro
for host headers. This will be resolved in other commits.

Signed-off-by: Artur Gainullin <artur.gainullin@intel.com>