[SYCL] Fix bug with not found alloca command.
Fix bug with not found alloca in case of two accessors are created for
the same memory object. The problem is that during searching for
dependencies for the first accessor list of leaf command is modified so
it doesn't contain commands that are considered as dependencies. As a
result search for dependencies for the second accessor fails.
The fix is to update list of leafs after dependencies for all accessors
are found.

Signed-off-by: Vlad Romanov <vlad.romanov@intel.com>