[SYCL] Document the SYCL Runtime Plugin Interface.
Initial version of the SYCL runtime plugin interface document.
The SYCL Runtime Plugin Interface (PI) is the interface layer between
device-agnostic part of the SYCL runtime and the device-specific runtime layers
which control execution on the device. It employs the plugin mechanism to bind
to the device specific runtime layers similarly to what is used by libomptarget
or OpenCL.

Signed-off-by: Konstantin Bobrovsky konstantin.s.bobrovsky@intel.com