[SYCL] A fix for a redefinition in nan builtin helper on MSVC compiler
Before this patch the helper did explicitly map integer types to
suitable floating point type.
aka
  template<> struct helper<cl_int> { using type = cl_int;}
But this does not work on Windows compiler x64 because there is
cl_long defined as an alias of long long. And this helper did generate
the redefinition error.

Fixed by change the mapping rules of the helper to mapping by size
comparing of integer and floating point types.
aka
  template<T>
  struct helper<T, enable_if_t<sizeof(T) == sizeof(cl_float), T>>
  { using type = cl_float; }

Signed-off-by: Alexey Voronov <alexey.voronov@intel.com>