[SYCL][Driver][Offloader] Adding appropriate error out on incorrect SYCL target
No longer will the switch statement assert with llvm_unreachable, 
instead the appropriate Clang diagnostic warning for an invalid SYCL 
output should be emitted. This occurs in the initialize() call for the 
SYCL offloader/phase and action generator, similar to how CUDA emits 
errors for incorrect GPU Arch targets.

Doing it similarly to OpenMP (inside CompilerInvocation.cpp) doesn't 
work in this case as if we pass an invalid target we won't accidentally 
generate an invalid/incorrect ToolChain like GetToolChain's we will 
instead return a nullptr which will ICE the compiler before it reaches 
CompilerInvocation.cpp, so the warning interaction has to be slightly 
different in this case.

Also removing some unused variables I missed in the first commit.

Signed-off-by: Andrew Gozillon <andrew.gozillon@yahoo.com>