[SYCL] Return __constant pointers from 'constant' multi_ptr
We cannot return a generic pointer like in 'global', 'local' and
'private' cases, because 'constant' and 'generic' address spaces do
not overlap.

This patch returns __constant pointer as is, and assumes that a user
have to deal with __constant pointer limitations (ie. such pointers
cannot be casted to plain (default) pointers).

Signed-off-by: Andrew Savonichev <andrew.savonichev@intel.com>