[SYCL] Turn new address space rules by default
This patch enables new rules for address space handling in SYCL by
default. Previous rules for address space handling can still be
enabled by DISABLE_INFER_AS=1 environment variable.

This approach is known as "generic pointers by default", and
it essentially makes all pointers without an explicit address space
qualifier to be pointers in generic address space.

For example:

    void foo() {
      int* ip = foo();
    }

    void* bar(int* ip);

    template <typename T>
    void swap(T&, T&) {}

All pointers (and references) in the example above are CodeGen'ed as
pointers in addrspace(4) in LLVM IR (where '4' is a number for generic
address space).

Concrete address space (global, local and private) can originate from:

  1) an address-of operation applied to an automatic variable:
     int i = 42;
     (void) &i // <--- this is a private pointer

  2) Static variable

  3) SYCL global, local, or constant buffers

In all three cases, once this pointer is used, it is immediately
addrspacecast'ed to generic, because a user does not (and should not)
specify address space for pointers in source code.

Signed-off-by: Andrew Savonichev <andrew.savonichev@intel.com>