[SYCL] Fix SFINAE for accessor class to work with MSVC
It looks like in the following case:
template<int T>
some_class_method(typename std::enable_if<[some type of the argument],
[some condition that doesn't depend on T] && T != 1>::type Arg) {
  ...
}

MSVC emits the error:
'type': is not a member of of any direct or indirect base class of
'std::enable_if<false,...'

This could be work arounded by changing the order of && arguments.
Probably the reason is that of first argument of && is false then we
don't even take into account the second argument.

Signed-off-by: Artur Gainullin <artur.gainullin@intel.com>