[FPGA] Implemented intel __fpga_reg built-in
Add a new compiler built-in function to enable low level optimization on
Intel FPGA device family. It should be implemented as an assignment
within the emulator.

Motivation: The function is __builtin_intel_fpga_reg( operand  ), and is legal
only in the kernel language. It takes a single argument/operand which can be
a first class lvalue (not a channel for example) or an rvalue.
In FPGA hardware, it semantically means that at least one hardware register
will be inserted between the operand the output of the built-in. The effect
of this is observable only through the frequency that a design operates at
in FPGA hardware, so it lowers to an assignment in all other contexts.
It is safe to optimize away during CPU compilation. Only very advanced users
should make use of it.

Implementation strategy: In CodeGen we generate one of two intrinsics:
* T llvm.annotation.*(T, '__builtin_intel_fpga_reg', ...) - it used for
                                           all types except structures/unions.
* T* llvm.ptr.annotation.*(T*, '__builtin_intel_fpga_reg', ...) - it used for
                                           structures/unions.

Signed-off-by: Mikhail Lychkov <mikhail.lychkov@intel.com>