[SYCL] Enable programs that include system headers on Windows
Add MSVC-specific predefined type characteristics for wchar, builtin
VaListKind, and a temporary workaround for the vectorcall attribute.
We'll allow vectorcall to appear on function declarations, since they
occur in the Microsoft headers, but we won't allow or will ignore
vectorcall on function invocations from SYCL target. The handling for
vectorcall is a work in progress.

Signed-off-by: Melanie Blower <melanie.blower@intel.com>