[SYCL] Fix implementation of vec class for half
Currently on the host side OpenCL cl_half# types are used as an
underlying type for vec class. But they are defined as an integer type
under the hood. As a result half values are converted to the integer
and passed to the kernel which expects a floating point number. This
patch fixes this bug.

Signed-off-by: Artur Gainullin <artur.gainullin@intel.com>