[SYCL] Add -fsycl-unnamed-lambda flag
User activates the extension with the new flag rather
than previously with -DUNNAMED_LAMBDA_EXT.

Fixes that the err_sycl_kernel_name_class_not_top_level
is an error without the extension and not a warning with it.

Only the version of the integration header needed is generated
(rather than both version with #ifdef).

Defines __SYCL_UNNAMED_LAMBDA__ when extension is active.

Signed-off-by: Roland Schulz <roland.schulz@intel.com>