[SYCL] Fix vector class relational/logical operations for length 1
When such an operator is called for a 1-element vector, the latter
is implicitly casted to a scalar, and only then is the operation
performed. Thereby, a scalar 0/1 logic is followed for a vector case,
which requires 0/-1 return value instead.

In future, SYCL may have to enforce scalar behaviour for 1-element vector,
thus avoiding any ambiguities.

Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>

Co-authored by Alexey Bader <alexey.bader@intel.com>