[SYCL] Add support for archives with fat objects (#384)
There are two parts in this commit.

First adds support for unbundling static libraries (or archives) with fat objects to clang-offload-bundler tool under a new type "ao". Output file for this type contains a list of temporary files containing device objects extracted from the archive's objects. Bundling operation for "ao" type is not supported.

And the second part contains driver changes for enabling -foffload-static-lib=... option in MSVC environment. Driver unbundles libraries specified in this option (output is a list file with extracted target object) and adds output target list files to the appropriate device link command as input. On the host side offload libraries are added to the link command as is ignoring output from the unbundling action.

Signed-off-by: Sergey Dmitriev <serguei.n.dmitriev@intel.com>