[SYCL] Use initialization instead of assignment for kernel objects
Changed OpenCL kernel generation to use initialization of the kernel
object instead of doing assignment.
The Sema routines (as well as CodeGen) expect a legal C++ AST.
Lvalue->Rvalue casts of structs and assignment to const fields is not legal C++.
Using assignment to initialize class objects also is not legal C++ as well since
a constructor should be called.

Signed-off-by: Mariya Podchishchaeva <mariya.podchishchaeva@intel.com>