[SYCL] Throw if loadProgram of SPIR-V/BC and no compiler available
is_compiler_available may be false if there is no JIT compiler
available.  One example is an FPGA.  If loadProgram finds a SPIR-V or
LLVM BC image, it will eventually call createProgramWithIL, which may
not be supported on a device.

Check that at least one Device in the Context supports
is_compiler_available.  If not, throw a feature_not_supported
exception.

Signed-off-by: Mark Mendel <mark.p.mendell@intel.com>