[SYCL] Fixes to make program_manager/program_manager.cpp pass on Windows
There were two issues which prevented test from running. First problem
was in the OSUtil::getOSModuleHandle(). It did not return the predefined
handle for the executable's module which in turn did not allow
program manager to find the target program that was registered by
__tgt_register_lib at startup. And the second problem was in the misuse
of std::vector<...>::reserve() for increasing vector size which lead to
a write to unallocated memory.

Signed-off-by: Sergey Dmitriev <serguei.n.dmitriev@intel.com>