[SYCL] Fix assertion for sycl keyword using without -fsycl option. (#445)
This is to fix assertion when clang compile C++ head file without -fsycl

The problem is when __pipe (one of sycl keyworld) used without
-fsycl
option, the __pipe should not be keyword and keyword status should be
KS_Disable. But it is wrongly set to KS_Future.

The root problem is the KEYSYCL is add to KEYALLCXX:

KEYALLCXX = KEYSYCL | KEYCXX | KEYCXX11 | KEYCXX2A

It seems, we want KEYSYCL to be part of feature of C++. But which C++
should belong: C++11 or C++2a? (for example KEYCX2A... vs concept)

For now we should just use -fsycl option to control this.

Signed-off-by: Yu jennifer.yu@intel.com