[SYCL] Add -reuse-exe to support FPGA recompile
Add -reuse-exe=<executable> option to support smart recompilation for
the Intel FPGA AOT compile.  This will extract the AOCX from the
executable and pass the extracted AOCX to the 'aoc' command line.
If the device (SPIR-V) code hasn't changed from the previous compile and
all options are the same, this will avoid re-running Quartus, which will
save a lot of compilation time when compiling for the FPGA.

If the executable does not exist, a warning is printed and it is
ignored. Add a new intel-fpga diagnostic group to allow the warning to
be disabled.

Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>