[SYCL] Emit extra pointer cast for @llvm.ptr.annotation
When a field of a C++ struct used in a constructor, `this' resides in
generic address space, and thus a pointer to the field also have
generic address space. With that saying we create llvm.ptr.annotation
intrinsic call and declaration with an argument with specified AS.

Also in this patch we ensure, that prior creating llvm.ptr.annotation call,
we are aligning type of the intrinsic's argument with i8 pointer in case
if argument is of not integer type and leaving it without changes if it is.

Signed-off-by: Dmitry Sidorov <dmitry.sidorov@intel.com>