[SYCL] Fix for memory layout and alignment of the class vec.
Now vectors with a length of 3 will store 4 elements of dataT.

Fixed boolean lit test on debug mode.
Fixed the alignment of vec<half, N> variants.

Now the vec class conform the requirements of item
"4.10.2.6 Memory layout and alignment" of the SYCL specification:
'''
The elements of an instance of the SYCL vec class template are stored
in memory sequentially and contiguously and are aligned to the size
of the element type in bytes multiplied by the number of elements:
  sizeof(dataT) * numElements
The exception to this is when the number of element is three in which
case the SYCL vec is aligned to the size of the element type in bytes
multiplied by four:
  sizeof(dataT) * 4.
'''

The lit test type.cpp was splited into two: aliases.cpp and types.cpp.
This is done to make the tests smaller and make the tests consistent
with what they're testing.

Signed-off-by: Alexey Voronov <alexey.voronov@intel.com>