[Clang][Bundler] Do not require host triple for extracting device bundles (#571)
* [Clang][Bundler] Do not require host triple for extracting device bundles

Bundler currently requires host triple to be provided no matter if you are performing bundling or unbundling, but for unbundling operation such requirement is too restrictive. You may for example want to examine device part of the object for a particular offload target, but you have to extract host part as well even though you do not need it. Host triple isn't really needed for unbundling, so this patch removes that requirement.

This is a cherry-pick of two llvm.org patches
  https://reviews.llvm.org/D66598 (prerequisite for the second patch)
  https://reviews.llvm.org/D66601

I addition to that this patch includes a fix for a bundler assertion on debug build

Signed-off-by: Sergey Dmitriev <serguei.n.dmitriev@intel.com>