[SYCL] Improve efficiency of stream output of h_item for FPGA.
This is done by creating a loop in writeHItem and ItemToStr,
rather than calling the same function multiple times.
Because everything is inlined in the FPGA flow, the latter results in many
copies of each function in hardware, which costs significant
FPGA area.  In a loop, the function will be created once on the
FPGA, and invoked multiple times.

Signed-off-by: David Cashman <david.cashman@intel.com>