[SYCL][Driver] Use of /Fo should not override intermediate offload files
When using clang-cl -fsycl, the addition of /Fo to name the output objects
would override the generation of expected temporary files in the offload
toolchain.

This allows for the temporary files to be used for the generated objects and
integrated header.  There is additional work that is needed to allow for
/Fofile to generate the object file along with the executable.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>