Add basic xmethod impl for accessor.operator[]
Intended to address templated function/operator debugging problem in
GDB. The added script is the currently used implementation in our gdb
- it only covers single-index `accessor.operator[]` with multiple
dimensions, that may be extended in the future depending on user
experience requirements.

It is much desirable to distribute this as part of compiler package
because it depends on private implementation details of SYCL headers and
must always be strictly in sync.

Signed-off-by: Mihails Strasuns <mihails.strasuns@intel.com>