[SYCL][Driver] file.lib files are not under consideration for unbundling
On windows, passing file.lib on the command line was triggering the use of the
unbundler.  Use of -foffload-static-lib is the way to pass libraries to use
for unbundling.  We key off of the file name extension to determine if an
object should be unbundled.  We need to do an explicit check for 'lib' with
the filename object check as there are some type lookups which rely on 'lib'
being an object

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>