[SYCL] Specify address space for pointer to data in accessor class
Accessor class contains pointer to data which is currently in the
generic address space. This causes SPIR-V Environment specification
violation for cases when accessor is wrapped to some object because
structure that has a pointer in the generic address space is not
allowed as a kernel argument. Fix specifies concrete address space for
this pointer in accessor class.

Signed-off-by: Artur Gainullin <artur.gainullin@intel.com>