[SYCL] Add entry table generation by wrapper for SYCL offload
This patch adds "-entries" option to clang-offload-wrapper which allows
to specify a file listing all function entries that are contained within
an image. This file is generated by the sycl-post-link tool and the entry
table will then be used in SYCL runtime to select a binary based on the
requested kernel.

Signed-off-by: Sergey Semenov <sergey.semenov@intel.com>