[SYCL] Add intel::kernel_args_restrict attribute
When applied to a device function that is invoked as a device kernel,
the attribute is a hint to the compiler that no pointer argument to
the kernel which is defined through an accessor (not USM), will alias
any other pointer kernel argument that was defined through an
accessor. This effect is equivalent to annotating restrict on all
kernel pointer arguments in an OpenCL or SPIR-V kernel.

Signed-off-by: Andrew Savonichev <andrew.savonichev@intel.com>