[SYCL] Change the way of accessing single element in sycl::vec
Returning scalars from sycl::vec class as a swizzled vec causes problems
with template resolution and ambiguity in overload resolution. This
commit implements workaround for this problem. Scalars are returned by
value for const vec and by reference for vec. Assumed that it is
possible to get a reference the element of the vector, test is added to
check that behaviour is not changed.

Signed-off-by: Artur Gainullin <artur.gainullin@intel.com>