[SYCL][Driver] Update -fintelfpga -fsycl-link behavior for unbundling objects (#777)
* [SYCL][Driver] Update -fintelfpga -fsycl-link behavior for unbundling objects

When unbundling objects when using -fsycl-link -fintelfpga, the expectation is
for the host and device parts of the object to be re-inserted into the FPGA
final device archive.  Updates to unbundling behavior causes only the host
objects to be extracted during the ubundling phase (this used to just return
the full fat object).  Now make sure the full object is passed to the archive
step.

* [NFC] Improvements via new boolean and update comment

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>