[SYCL][Driver] Use original archive when linking with an FPGA AOCX binary
When performing a final link step with -fintelfpga and given an AOCX device
archive, use the archive to link with as opposed to the unbundled host
objects.  The recent changes to the unbundler does not return the full
fat object, causing the host objects to not be 'found' during the final host
unbundling.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>