[SYCL][Driver] Push linker input to the linker and not bundler when -foffload-static-lib is supplied
This fixes a problem that occurs when linker input like '-z relro' (and
some libraries that are translated to options internally e.g. -lstdc++)
are passed to the compiler alongside the -foffload-static-lib option.

Previously linker arguments were being pushed to the OffloadBundler's
ConstructJobMultipleOutputs call, which then got pushed into the linkers
job construction as a string. As they're not considered strings but an
Input (llvm::Opt) it will result in some garbage data as an argument to
the linker. This will cause the linker job before the OffloadBundler job
to fail as it can't find the garbage argument and then kills the rest of
the compilation e.g:

/usr/bin/ld: cannot find �H�: No such file or directory
clang-10: error: clang-offload-bundler command failed with exit code 1
(use -v to see invocation)

Signed-off-by: Andrew Gozillon andrew.gozillon@yahoo.com