[SYCL] Handle captures by ref in kernel lambda
Captures by reference wasn't properly handled
by the compiler which led to a crash.

Values captured by ref are not standard layout
so if we see one, emit a corresponding error
as for other non-standard layout structures/classes.

This patch also slightly changes diagnostic for
non-standard layout types: in case of explicit
capture point to the capture location, otherwise
point to the first usage of a captured variable
inside a lambda. This should be more informative
because standard layout property of a capture inside lambda
may not match the one outside(especially for capture by ref).

Signed-off-by: Ilya Stepykin <ilya.stepykin@intel.com>