[SYCL][Driver] fix -fsycl -lstdc++ abort problem
Use of -fsycl -lstdc++ causes an abort when processing Input arguments
slated for unbundling.  -lstdc++ is a special case library that is transformed
to a different value.  This new input arg ends up having no actual value
negating the assumption that Input args have a value.  Any LinkerInput
options/Args are not to be considered to be unbundled.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>