[SYCL] Implement IVDep for arrays, dependent cases, general cases (#779)
IVDep marks loops and arrays in a loop with a safe length. Metadata gets
applied to each array access GEP to label the safe length for that
array.

Additionally, this patch works with dependent arguments and adds the
infrastructure to do so.  This functionality doesn't currently exist for
statement attributes.

Signed-off-by: Erich Keane <erich.keane@intel.com>