[SYCL][Driver] project report output not correct for Windows FPGA AOT
When building for Windows FPGA AOT, the use of -o was not being applied
to the setting of the output report project location.  Make adjustments
to allow for the Windows -o to apply as well as use /Fe, which is the
more natural output setting usage on Windows.

Also, update the test to use the proper clangxx instead of clang++

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>