[SYCL][Scheduler] Refactor data transfer scheduler commands
The patch makes transfer commands such as map and unmap more flexible.
Now they can be used not only for situations when we have host accessor
which we need to initialize with host ptr, but for doing map/unmap in
cases when we can/must avoid copying which would require additional
memory allocation on the host.

Signed-off-by: Vlad Romanov <vlad.romanov@intel.com>