[SYCL] Add SYCLConfig helper
SYCLConfig helper provides an access to configs that is defined in
config.hpp. Such configs can be set at compile time, using environment
variable and configuration file. SYCLConfig checks all the sources and
return value which was set using an option which has highest priority.
Current priorities are:
1. Environment variable
2. Configuration file
3. Compiler time

If a value is not set nullptr is returned.

Signed-off-by: Vlad Romanov <vlad.romanov@intel.com>