[SYCL] Fix compile error for handler::copy with -fsycl-unnamed-lambda
Some `copy` methods in handler are implemented via `parallel_for` for
host device and hence, we need to have `KernelInfo` struct (provided by
integration header) to compile them.
This effectively creates cyclic dependency and we already have some kind
of a workaround for it for regular mode (without -fsycl-unnamed-lambda).

This patch applies the same solution for unnamed lambda mechanism.

Signed-off-by: Alexey Sachkov <alexey.sachkov@intel.com>