[SYCL] Allow SYCL_EXTERNAL to be applied to a function with raw pointers (#861)
SYCL 1.2.1 specification does not allow SYCL_EXTERNAL to be applied to
a function with a pointer argument or with a pointer return
value. This is required to limit address space inference work to a
single translation unit.

Given that we make inference after we link all translation units
together, it should not really matter now. And there are several cases
where external functions with pointers can be useful, so this patch
tuns an error into a warning.

The default behavior is to report an error. `-Wno-error=sycl-strict` allows
raw pointer function parameters or return type.

Signed-off-by: Andrew Savonichev <andrew.savonichev@intel.com>