[Driver][NFC] cleanup unused -fsycl-link FPGA bundling settings
When performing -fsycl-link -fintelfpga, there is no bundling step that occurs
which uses the FPGA type specific settings (aocr, aocx, aoco).  If a bundling
step does occur (with -c), it is only for the specific device compile and does
not include the AOT step.

These settings were part of an earlier implementation that considered using
a bundled binary instead of an archive to store the aocx/aocr generated bits

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>