[SYCL][Driver] Fix output options behavior for -fsycl-link on Windows
When specifying output options with -fsycl-link -fintelfpga on Windows, the
named output is not used.  This is due to the output type for -fsycl-link
being an archive type not being recognized when generating output file names.

Also address a problem seen where use of /Fo was using that name for the
intermediate unbundling step.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>