[SYCL] Fix regression introduced by f537293/part of #1018 fix.
This fixes the immediate problem, where diagnostics were being missed
in kernels.  The problem ends up being that diagnostics can happen
at three different times during translation for templates:

1- During Phase 1
2- During Phase 2 instantiation
3- During ConstructOpenCLKernels.

f537293 suppressed 3, and did a better job at suppressing 1. However,
CallExprs are always rebuilt during instantation (for technical
reasons).  The result was the variadic call was being diagnosed in all 3
while developing f537293.  I erronously thought the best way was to
simply wait for instantiation to diagnose all SYCL diagnostics.

Unfortunately, this was short-sighted. Not everything in tree-transform
gets rebuilt every time.  One perfect example is the inline-asm (see the
added example). It doesn't get rebuilt in some cases, so it was never
diagnosed during instantiation.  Because I had fixed Phase
1/ConstructOpenCLKernels, this left no time for it to be diagnosed!

The fix was to keep suppressing 3, and simply fix the call-expr check by
moving the check to Sema::checkCall. This is how Clang works around this
issue for warnings. The downside is that the diagnostics will happen
slightly later for this (see the sycl-restrict.cpp test, which reverted
to the previous version before f537293), but invalid programs still
won't be accepted.

Signed-off-by: Erich Keane <erich.keane@intel.com>