[SYCL][FPGA] Enable uses_global_work_offset attribute (#1010)
`uses_global_work_offset` attribute applies to a device function/lambda function or function call operator. Valid values are 0 and 1.
If value >1 is specified as an argument, we warn user that it will be adjusted to 1. As attribute is not ignored in this case and it does not match any existing diagnostic group, the new `AdjustedAttributes` group was added.

Signed-off-by: Viktoria Maksimova <viktoria.maksimova@intel.com>